Add missing test for the dashboard controller
