+++ test/crawlexir_web/controllers/upload_controller_test.exs
@@ -1,6 +1,5 @@
 defmodule CrawlexirWeb.UploadControllerTest do
   use CrawlexirWeb.ConnCase
-  use CrawlexirWeb.ControllerCase
 
   alias Crawlexir.Search
